+++ tools/docmissing.py
@@ -11,6 +11,33 @@ import apsw
 
 import names
 
+# check shell knows all pragmas
+import apsw.shell
+
+con = apsw.Connection("")
+all_pragmas = set(con.execute("pragma pragma_list").get)
+deprecated_pragmas = {
+    "count_changes",
+    "empty_result_callbacks",
+    "full_column_names",
+    "legacy_file_format",
+    "short_column_names",
+    "temp_store_directory",
+}
+for pragma in all_pragmas:
+    if pragma in deprecated_pragmas:
+        continue
+    check = (pragma, f"{pragma}=", f"{pragma}(", f"{pragma};")
+    assert any(c in apsw.shell.Shell._pragmas for c in check), f"pragma { pragma } not in apsw.shell.Shell._pragmas"
+
+# check all pragmas are known to sqlite
+for pragma in apsw.shell.Shell._pragmas:
+    for c in "=(;":
+        if pragma.endswith(c):
+            pragma = pragma[:-1]
+    assert pragma in all_pragmas or pragma in deprecated_pragmas, f"{pragma} is in shell but not known to SQLite"
+
+
 retval = 0
 
 classes = {}
@@ -38,8 +65,9 @@ cur = con.cursor()
 cur.execute("create table x(y); insert into x values(x'abcdef1012');select * from x")
 blob = con.blob_open("main", "x", "y", con.last_insert_rowid(), 0)
 vfs = apsw.VFS("aname", "")
-vfsfile = apsw.VFSFile("", con.db_filename("main"),
-                       [apsw.SQLITE_OPEN_MAIN_DB | apsw.SQLITE_OPEN_CREATE | apsw.SQLITE_OPEN_READWRITE, 0])
+vfsfile = apsw.VFSFile(
+    "", con.db_filename("main"), [apsw.SQLITE_OPEN_MAIN_DB | apsw.SQLITE_OPEN_CREATE | apsw.SQLITE_OPEN_READWRITE, 0]
+)
 
 # virtual tables aren't real - just check their size hasn't changed
 for n, e in (("VTModule", 3), ("VTTable", 17), ("VTCursor", 7)):
@@ -48,13 +76,13 @@ for n, e in (("VTModule", 3), ("VTTable", 17), ("VTCursor", 7)):
     del classes[n]
 
 for name, obj in (
-    ('Connection', con),
-    ('Cursor', cur),
-    ('Blob', blob),
-    ('VFS', vfs),
-    ('VFSFile', vfsfile),
-    ('apsw', apsw),
-    ('VFSFcntlPragma', apsw.VFSFcntlPragma),
+    ("Connection", con),
+    ("Cursor", cur),
+    ("Blob", blob),
+    ("VFS", vfs),
+    ("VFSFile", vfsfile),
+    ("apsw", apsw),
+    ("VFSFcntlPragma", apsw.VFSFcntlPragma),
     ("zeroblob", apsw.zeroblob(3)),
 ):
     if name not in classes:
@@ -67,7 +95,8 @@ for name, obj in (
             retval = 1
             print("%s.%s in documentation but not object" % (name, c))
     for c in dir(obj):
-        if c.startswith("__"): continue
+        if c.startswith("__"):
+            continue
         # old renamed names?
         if name in names.renames and c in names.renames[name].values():
             continue
@@ -85,13 +114,26 @@ for name, obj in (
             if isinstance(getattr(apsw, c), type) and issubclass(getattr(apsw, c), Exception):
                 continue
             # ignore classes !!!
-            if c in ("Connection", "VFS", "VFSFile", "zeroblob", "Shell", "URIFilename", "Cursor", "Blob", "Backup",
-                     "IndexInfo", "VFSFcntlPragma", "FTS5Tokenizer", "FTS5ExtensionApi"):
+            if c in (
+                "Connection",
+                "VFS",
+                "VFSFile",
+                "zeroblob",
+                "Shell",
+                "URIFilename",
+                "Cursor",
+                "Blob",
+                "Backup",
+                "IndexInfo",
+                "VFSFcntlPragma",
+                "FTS5Tokenizer",
+                "FTS5ExtensionApi",
+            ):
                 continue
             # ignore mappings !!!
             if c.startswith("mapping_"):
                 continue
-        if c not in classes[name]:
+        if c not in classes[name] and not c.startswith("_"):
             retval = 1
             print("%s.%s on object but not in documentation" % (name, c))