+++ tools/genfaultinject.py
@@ -92,6 +92,10 @@ returns = {
             PyMem_Realloc PyUnicode_FromFormat PyTuple_Pack PyUnicode_AsUTF8String
             PyObject_VectorcallMethod PyObject_Vectorcall PyWeakref_GetObject PyUnicode_DecodeUTF8
             Connection_fts5_api get_token_value fts5extensionapi_acquire
+
+            sqlite3_malloc sqlite3_malloc64 sqlite3_mprintf
+            sqlite3_realloc sqlite3_realloc64
+            sqlite3_normalized_sql sqlite3_expanded_sql
             """.split(),
     # numeric return
     "sqlite":
@@ -116,12 +120,12 @@ returns = {
             sqlite3_declare_vtab sqlite3_deserialize
             sqlite3_drop_modules sqlite3_enable_load_extension
             sqlite3_enable_shared_cache sqlite3_exec
-            sqlite3_expanded_sql sqlite3_initialize
-            sqlite3_load_extension sqlite3_malloc sqlite3_malloc64
-            sqlite3_mprintf sqlite3_normalized_sql sqlite3_open
+            sqlite3_initialize
+            sqlite3_load_extension
+            sqlite3_open
             sqlite3_open_v2 sqlite3_overload_function
-            sqlite3_prepare_v3 sqlite3_realloc
-            sqlite3_realloc64 sqlite3_result_zeroblob64
+            sqlite3_prepare_v3
+            sqlite3_result_zeroblob64
             sqlite3_set_authorizer sqlite3_shutdown sqlite3_status64
             sqlite3_table_column_metadata sqlite3_threadsafe
             sqlite3_trace_v2 sqlite3_vfs_register